backend/requirements.txt
# Backend dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
deep-translator>=1.11.4
mangum>=0.17.0
boto3>=1.34.0
